requirements.txt
opencv-python-headless
pytesseract
flask
numpy
Pillow
requests
